+++ reactivestate-core/src/commonMain/kotlin/com/ensody/reactivestate/Utils.kt
@@ -22,7 +22,7 @@ private class LazyProperty<T>(block: (property: KProperty<*>) -> T) : ReadOnlyPr
 }
 
 /** Wraps a value. Together with nullability can model an `Option`/`Maybe`. */
-public data class Wrapped<T>(public val value: T) : Serializable, ReadOnlyProperty<Any?, T> {
+public data class Wrapped<T>(public val value: T) : JvmSerializable, ReadOnlyProperty<Any?, T> {
     override fun getValue(thisRef: Any?, property: KProperty<*>): T = value
 
     override fun toString(): String =