+++ reactivestate-core/src/nonJvmMain/kotlin/com/ensody/reactivestate/Serializable.nonJvm.kt
@@ -1,3 +1,6 @@
 package com.ensody.reactivestate
 
-public actual interface Serializable
+public actual interface JvmSerializable
+
+public actual fun <T : Any> JvmSerializerReplacement(serializer: RawSerializer<T>, value: T): Any =
+    DummyJvmSimpleSerializerReplacement